scrapperYT/requirements.txt
pymongo
redis
python-dotenv
schedule
requests